+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## NEXT RELEASE
 ### Added
 - Added config `sitemap.route` to enable a route delivering a dynamic XML-sitemap.
+- Added config `no_prefix_locales` to disable locale-prefix for specific languages.
 
 ## [v2.2.1 (2020-03-04)](https://github.com/webflorist/routetree/compare/v2.2.0...v2.2.1)
 ### Added